pyproject.toml
[tool.poetry]
name = "intelli-trip-planner"
version = "0.0.2"
description = "[https://github.com/BadwaterBay/intelli-trip-planner](https://github.com/BadwaterBay/intelli-trip-planner)"
authors = ["Badwater Bay <info@badwaterbay.com>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = "~3.8"
Flask = "^1.1.2"
ortools = "^7.8.7959"
googlemaps = "^4.4.2"
python-dotenv = "^0.14.0"
gunicorn = "^20.0.4"
coverage = "^5.2.1"
httpx = "^0.14.3"
fuzzywuzzy = "^0.18.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
black = "^20.8b1"
coverage = "^5.2.1"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
